refactor: move the tools-pane and repo routes out of index.ts (#548 step 2c-2)
Six routes, two groups. The tools pane's four — the toolResult sink, its replay,
the available-tool list and the call history — plus /prs and /issues, which only
ever needed config and two git helpers.

The stores go in as one object rather than four functions: they are a single
instance owned by index.ts, and handing the routes their own would give each an
in-memory copy of the same sessions. `publish` is again a closure, for the same
reason as the stores themselves.

Verified against main side by side, with the validation paths included since
they moved too: eleven requests — a stored result read back, a persistOnly
update that must not re-publish, the three rejected shapes (bad session id,
missing toolName, missing uuid), and both repo listings — identical.

index.ts: 2223 → 2146 lines; inline routes 14 → 8.

Refs #548
